Qualif/v2.4.5 rc1

+++ ci/deploy.sh
@@ -81,7 +81,7 @@ if [ -z "$HEROKU_APP" ] ;then
   exit 1
 fi
 
-export NETLIFY_SITE_ID=$HEROKU_APP.netlify.com
+export NETLIFY_SITE_ID=${NETLIFY_SITE_ID:-$HEROKU_APP.netlify.com}
 
 #
 # Deploy API on heroku
@@ -129,9 +129,13 @@ heroku container:release web --app $HEROKU_APP
 
 # first time : setup DB
 BRANCH_DIR=$(echo "$BRANCH" |tr '/' '-')
-heroku run --no-tty -a $HEROKU_APP "( cd /app && wget  -L https://github.com/LAB-MI/candilibV2/archive/${BRANCH}.tar.gz -O - |tar zxvf - --strip-components=2 candilibV2-${BRANCH_DIR}/server/dev-setup ) && npm run dev-setup"
+GIT_REMOTE_URL=$(git config --get remote.origin.url | sed -e 's/.git$//g')
+heroku run --no-tty -a $HEROKU_APP "set -e ; ( cd /app && wget  -L ${GIT_REMOTE_URL}/archive/${BRANCH}.tar.gz -O - |tar zxvf - --strip-components=2 candilibV2-${BRANCH_DIR}/server/dev-setup ) && npm run dev-setup"
+res=$?
 # TODO: variable REPO
-
+if  [ "$res" -gt 0 ] ; then
+   exit $res
+fi
 rm -rf ~/.netrc
 fi
 
@@ -153,13 +157,14 @@ fi
 # Install netlify-cli
 curl -sL https://deb.nodesource.com/setup_12.x | sudo -E bash -
 sudo apt-get install -y nodejs
+node --version
 npm --version
-sudo npm install -g netlify-cli
+sudo npm install -g netlify-cli@2.24.0
 
 # Extract front files from previously docker images
 make export-front-all
-APP_VERSION="$(./ci/version.sh)"
 
+APP_VERSION="$(./ci/version.sh)"
 (
 cd "candilibV2-$APP_VERSION-dist"
 
@@ -183,10 +188,15 @@ cat > dist/index.html <<EOF
 EOF
 
 # deploy files
-netlify deploy --prod --dir=dist --message "Deploy $APP_VERSION" --json | tee -a netlify.json
+json=$(netlify deploy --timeout 30 --prod --dir=./dist --message "Deploy $APP_VERSION" --json)
+echo "$json" > netlify.json
+cat netlify.json
 
 deploy_id="$(jq -re '.deploy_id' < netlify.json)"
-
+if [ "$?" -gt 0 ] || [ -z "$deploy_id" ] ; then
+   echo "Netlify error: deploy_id empty"
+   exit 1
+fi
 curl_args="--fail --retry 10 --retry-delay 5 --retry-max-time 60 --connect-timeout 10 "
 # lock deployment
 curl ${curl_args} -o /dev/null -X POST -H "Content-Type: application/json" \
